Clean up inconsistent code style.
I've waffled over time about code style in Maya plugins, and some of them use camelCase like
the Maya API, but the right thing to do is to use standard Python code style wherever possible
and treat the Maya API as the weird exception, since that's what it is.

Clean some of this up, since it's resulted in mixed code styles.