Rename zSimpleRigging to zMayaUtils.
Helpers that don't add Maya nodes and which don't do things that people might
not want (like zShelfMenu) will go in here, so we don't end up with a ton of
modules that need to be loaded.  This doesn't have any real cost if it's not used,
and by only importing modules in the command handler we can avoid pulling in
dependencies that might not always be there, like the native OBB plugin.