Remove mirror weights.
This oesn't offer much that the builtin mirror deformer weights offers, and I
didn't see that it was there since it's in a weird place in the menu, and since
copy weights doesn't exist I probably assumed mirror weights didn't either.

It does allow mirroring blend shape weights, which the builtin one doesn't, but
it's not worth keeping around just for that.