Work on blend shape retargetting:
- Copy the shape editor group hierarchy, so retargetting a lot of blend shapes doesn't
reduce everything to a flat list.
- Use maya_helpers.SetAndRestore restores to make sure the source blend shape is enabled
and restore any changes when we're done.

This lets us enable the source blend shape directory, so we no longer need to make a
copy of the source mesh and blend shapes.  This makes retargetting much faster, and also
makes it work better if the source mesh has been modified by deformers (eg. rescaling
with a skinCluster).
- If "Connect weights to source" is turned off, copy the source weight to the new blend
shape instead of leaving everything at 0.
- Remove an old workaround for wrap.  Just set render influences to true.