Move shelf menu configuration to preferences.
This is more natural than doing it within the shelf menu.

This also now allows hiding the top-level "Shelves" menu, which is now off
by default.  It hasn't been very useful, and adding an unimportant top-level
menu whenever zMayaUtils is loaded just clutters the menu bar.  This can be
turned back on in preferences.