Rework EyeRig.
Derive the rotation from the aim constraint we were using to orient the
handle.  This takes significantly fewer nodes, and fixes eyesFocused not
being composited correctly.

The angleLeft and angleRight attributes are removed.  They should always
be the same as the X and Y rotations of the output joints.

This now requires Maya 2020.