Encase the paths used for mkdir in quotation marks
I ran into an issue with the installer where OneDrive had formatted the name of my OneDrive directory as "OneDrive - CompanyName". The - surrounded by spaces made the path evaluate as an argument when passed to mkdir and caused the script to fail. 
Encasing the variables used by mkdir should ensure the whole path gets evaluated as a string. I made this simple change and it worked without problem.